# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger ManhDanBlogs",
        "description": "This is a sample Userstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://lialili.fly.dev/api/",
            "description": "Development Environment"
        },
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Staging  Environment"
        }
    ],
    "paths": {
...
        "/admin/product/getProductImage/{product_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Product Image",
                "operationId": "adgetProductImage",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product Image ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/item/getProductImage/{product_id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get Product Image",
                "operationId": "getProductImage",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product Image ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "required": [
                    "id",
                    "title",
                    "category_id",
                    "slug",
                    "prince",
                    "description",
                    "image"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "description": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}